feat: redirect if user cannot upload
